Correctly Create and Return Reviews -change id values from long to int -return reviews with beer

+++ HopHubApi/Repositories/IBeerRepository.cs
@@ -7,7 +7,7 @@ namespace HopHubApi.Repositories
     public interface IBeerRepository
     {
         Task<List<Beer>> GetAllAsync();
-        Task<Beer> GetByIdAsync(long id);
+        Task<Beer> GetByIdAsync(int id);
         Task CreateAsync(Beer beer);
         Task UpdateAsync(Beer beer);
         Task DeleteAsync(Beer beer);
